requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-telegram-bot==20.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
